INS-3143 move some BadgerTests to slow build tag
