+++ ledger/drop/drop_cmp_slow_test.go
@@ -1,0 +1,116 @@
+//
+// Copyright 2019 Insolar Technologies GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+// +build slowtest
+
+package drop
+
+import (
+	"context"
+	"io/ioutil"
+	"os"
+	"testing"
+
+	fuzz "github.com/google/gofuzz"
+	"github.com/insolar/insolar/insolar/gen"
+	"github.com/insolar/insolar/insolar/store"
+	"github.com/insolar/insolar/instrumentation/inslogger"
+	"github.com/stretchr/testify/require"
+)
+
+func TestDropStorageDB(t *testing.T) {
+	ctx := inslogger.TestContext(t)
+	tmpdir, err := ioutil.TempDir("", "bdb-test-")
+	defer os.RemoveAll(tmpdir)
+	require.NoError(t, err)
+
+	db, err := store.NewBadgerDB(tmpdir)
+	require.NoError(t, err)
+	defer db.Stop(context.Background())
+	ds := NewDB(db)
+
+	var drops []Drop
+	genInputs := map[jetPulse]struct{}{}
+	f := fuzz.New().Funcs(func(jd *Drop, c fuzz.Continue) {
+		pn := gen.PulseNumber()
+		jd.Pulse = pn
+
+		jetID := gen.JetID()
+		jd.JetID = jetID
+
+		genInputs[jetPulse{jetID: jetID, pn: pn}] = struct{}{}
+	}).NumElements(5, 1000)
+	f.Fuzz(&drops)
+
+	// Add
+	for _, dr := range drops {
+		err := ds.Set(ctx, dr)
+		require.NoError(t, err)
+	}
+
+	// Fetch
+	for inp := range genInputs {
+		_, err := ds.ForPulse(ctx, inp.jetID, inp.pn)
+		require.NoError(t, err)
+	}
+}
+
+func TestDropStorageCompare(t *testing.T) {
+	ctx := inslogger.TestContext(t)
+
+	tmpdir, err := ioutil.TempDir("", "bdb-test-")
+	defer os.RemoveAll(tmpdir)
+	require.NoError(t, err)
+
+	db, err := store.NewBadgerDB(tmpdir)
+	require.NoError(t, err)
+	defer db.Stop(context.Background())
+	ds := NewDB(db)
+	ms := NewStorageMemory()
+
+	var drops []Drop
+	genInputs := map[jetPulse]struct{}{}
+	f := fuzz.New().Funcs(func(jd *Drop, c fuzz.Continue) {
+		pn := gen.PulseNumber()
+		jd.Pulse = pn
+
+		jetID := gen.JetID()
+		jd.JetID = jetID
+
+		genInputs[jetPulse{jetID: jetID, pn: pn}] = struct{}{}
+	}).NumElements(5, 1000)
+	f.Fuzz(&drops)
+
+	// Add
+	for _, dr := range drops {
+		err := ds.Set(ctx, dr)
+		require.NoError(t, err)
+		err = ms.Set(ctx, dr)
+		require.NoError(t, err)
+	}
+
+	// Fetch
+	for inp := range genInputs {
+		dbDrop, err := ds.ForPulse(ctx, inp.jetID, inp.pn)
+		require.NoError(t, err)
+
+		memDrop, err := ms.ForPulse(ctx, inp.jetID, inp.pn)
+		require.NoError(t, err)
+
+		require.Equal(t, dbDrop, memDrop)
+	}
+
+}
